MULE-18492: Classloader for policy plugin used for serialization is not the actual classloader (#8925)
(cherry picked from commit 7813d87c167c2adb2128417bcb387dca78f98b4e)
