+++ modules/deployment/src/test/java/org/mule/runtime/module/deployment/internal/ApplicationDeploymentTestCase.java
@@ -7,20 +7,14 @@
 
 package org.mule.runtime.module.deployment.internal;
 
-import static java.io.File.separator;
 import static java.lang.String.format;
 import static java.lang.System.currentTimeMillis;
-import static java.lang.System.getProperty;
 import static java.lang.Thread.currentThread;
 import static java.util.Arrays.asList;
 import static java.util.Collections.emptyMap;
-import static org.apache.commons.io.FileUtils.copyFile;
 import static org.apache.commons.io.FileUtils.deleteDirectory;
-import static org.apache.commons.io.FileUtils.writeStringToFile;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.containsString;
 import static org.hamcrest.Matchers.equalTo;
-import static org.hamcrest.Matchers.instanceOf;
 import static org.hamcrest.Matchers.not;
 import static org.hamcrest.core.Is.is;
 import static org.hamcrest.core.IsNull.notNullValue;
@@ -38,11 +32,8 @@ import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mule.runtime.api.deployment.meta.Product.MULE;
-import static org.mule.runtime.container.api.MuleFoldersUtil.getMuleBaseFolder;
 import static org.mule.runtime.container.internal.ClasspathModuleDiscoverer.EXPORTED_CLASS_PACKAGES_PROPERTY;
 import static org.mule.runtime.container.internal.ClasspathModuleDiscoverer.EXPORTED_RESOURCE_PROPERTY;
-import static org.mule.runtime.container.internal.ClasspathModuleDiscoverer.PRIVILEGED_ARTIFACTS_PROPERTY;
-import static org.mule.runtime.container.internal.ClasspathModuleDiscoverer.PRIVILEGED_EXPORTED_CLASS_PACKAGES_PROPERTY;
 import static org.mule.runtime.core.api.config.MuleProperties.OBJECT_MULE_CONFIGURATION;
 import static org.mule.runtime.core.internal.config.bootstrap.ClassLoaderRegistryBootstrapDiscoverer.BOOTSTRAP_PROPERTIES;
 import static org.mule.runtime.deployment.model.api.DeployableArtifactDescriptor.PROPERTY_CONFIG_RESOURCES;
@@ -55,7 +46,6 @@ import static org.mule.runtime.extension.api.loader.xml.XmlExtensionModelLoader.
 import static org.mule.runtime.module.deployment.impl.internal.policy.PropertiesBundleDescriptorLoader.PROPERTIES_BUNDLE_DESCRIPTOR_LOADER_ID;
 import static org.mule.runtime.module.deployment.internal.DeploymentDirectoryWatcher.DEPLOYMENT_APPLICATION_PROPERTY;
 import static org.mule.runtime.module.deployment.internal.TestApplicationFactory.createTestApplicationFactory;
-import static org.mule.runtime.module.extension.api.loader.java.DefaultJavaExtensionModelLoader.JAVA_LOADER_ID;
 import static org.mule.tck.MuleTestUtils.testWithSystemProperty;
 
 import org.junit.Rule;
@@ -65,7 +55,6 @@ import org.mule.runtime.api.deployment.meta.MuleArtifactLoaderDescriptor;
 import org.mule.runtime.api.deployment.meta.MuleArtifactLoaderDescriptorBuilder;
 import org.mule.runtime.api.deployment.meta.MulePluginModel;
 import org.mule.runtime.api.exception.MuleException;
-import org.mule.runtime.api.exception.MuleFatalException;
 import org.mule.runtime.core.api.config.MuleConfiguration;
 import org.mule.runtime.core.api.config.MuleProperties;
 import org.mule.runtime.core.api.extension.ExtensionManager;
@@ -75,34 +64,23 @@ import org.mule.runtime.deployment.model.api.application.ApplicationStatus;
 import org.mule.runtime.deployment.model.internal.application.MuleApplicationClassLoaderFactory;
 import org.mule.runtime.deployment.model.internal.nativelib.DefaultNativeLibraryFinderFactory;
 import org.mule.runtime.extension.api.loader.xml.XmlExtensionModelLoader;
-import org.mule.runtime.module.artifact.builder.TestArtifactDescriptor;
 import org.mule.runtime.module.deployment.api.DeploymentListener;
 import org.mule.runtime.module.deployment.impl.internal.builder.ApplicationFileBuilder;
 import org.mule.runtime.module.deployment.impl.internal.builder.ArtifactPluginFileBuilder;
 import org.mule.runtime.module.deployment.impl.internal.builder.JarFileBuilder;
 import org.mule.runtime.module.deployment.impl.internal.domain.DefaultDomainManager;
 import org.mule.tck.junit4.rule.SystemProperty;
-import org.mule.tck.util.CompilerUtils;
-import org.mule.tck.util.CompilerUtils.JarCompiler;
-import org.mule.tck.util.CompilerUtils.SingleClassCompiler;
 
 import java.io.File;
 import java.net.URL;
-import java.nio.file.Paths;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Optional;
 import java.util.Properties;
-import java.util.Set;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.locks.ReentrantLock;
 
-import io.qameta.allure.Description;
-import io.qameta.allure.Flaky;
-import io.qameta.allure.Issue;
 import org.apache.commons.io.IOUtils;
 import org.junit.Before;
-import org.junit.Ignore;
 import org.junit.Test;
 
 /**
@@ -126,6 +104,7 @@ public class ApplicationDeploymentTestCase extends AbstractApplicationDeployment
     super(parallelDeployment);
   }
 
+  @Override
   @Before
   public void before() {
     incompleteAppFileBuilder = appFileBuilder("incomplete-app").definedBy("incomplete-app-config.xml");
@@ -379,30 +358,6 @@ public class ApplicationDeploymentTestCase extends AbstractApplicationDeployment
   }
 
   @Test
-  public void removesPreviousAppFolderOnRedeploy() throws Exception {
-    startDeployment();
-
-    addPackedAppFromBuilder(emptyAppFileBuilder);
-
-    assertApplicationDeploymentSuccess(applicationDeploymentListener, emptyAppFileBuilder.getId());
-
-    assertAppsDir(NONE, new String[] {emptyAppFileBuilder.getId()}, true);
-    assertEquals("Application has not been properly registered with Mule", 1, deploymentService.getApplications().size());
-
-    reset(applicationDeploymentListener);
-
-    ApplicationFileBuilder emptyAppFileBuilder =
-        appFileBuilder("empty-app").usingResource("empty-config.xml", "empty-config.xml")
-            .deployedWith(PROPERTY_CONFIG_RESOURCES, "empty-config.xml");
-
-    addPackedAppFromBuilder(emptyAppFileBuilder);
-
-    assertApplicationRedeploymentSuccess(emptyAppFileBuilder.getId());
-
-    assertApplicationFiles(emptyAppFileBuilder.getId(), new String[] {"empty-config.xml"});
-  }
-
-  @Test
   public void removesPreviousAppFolderOnStart() throws Exception {
     addExplodedAppFromBuilder(emptyAppFileBuilder);
 
@@ -420,26 +375,6 @@ public class ApplicationDeploymentTestCase extends AbstractApplicationDeployment
     assertAppsDir(NONE, new String[] {emptyAppFileBuilder.getId()}, true);
 
     assertApplicationFiles(emptyAppFileBuilder.getId(), new String[] {"empty-config.xml"});
-  }
-
-  @Test
-  public void redeploysAppZipDeployedAfterStartup() throws Exception {
-    addPackedAppFromBuilder(emptyAppFileBuilder);
-
-    startDeployment();
-
-    assertApplicationDeploymentSuccess(applicationDeploymentListener, emptyAppFileBuilder.getId());
-
-    assertAppsDir(NONE, new String[] {emptyAppFileBuilder.getId()}, true);
-    assertEquals("Application has not been properly registered with Mule", 1, deploymentService.getApplications().size());
-
-    reset(applicationDeploymentListener);
-
-    addPackedAppFromBuilder(emptyAppFileBuilder);
-
-    assertApplicationRedeploymentSuccess(emptyAppFileBuilder.getId());
-    assertEquals("Application has not been properly registered with Mule", 1, deploymentService.getApplications().size());
-    assertAppsDir(NONE, new String[] {emptyAppFileBuilder.getId()}, true);
   }
 
   @Test
@@ -554,214 +489,6 @@ public class ApplicationDeploymentTestCase extends AbstractApplicationDeployment
     assertAppsDir(NONE, new String[] {incompleteAppFileBuilder.getId()}, true);
     assertArtifactIsRegisteredAsZombie(incompleteAppFileBuilder.getId(), deploymentService.getZombieApplications());
     assertThat(deploymentService.findApplication(incompleteAppFileBuilder.getId()).getRegistry(), nullValue());
-  }
-
-  @Test
-  public void redeploysExplodedAppOnStartup() throws Exception {
-    addExplodedAppFromBuilder(dummyAppDescriptorFileBuilder);
-
-    startDeployment();
-
-    assertApplicationDeploymentSuccess(applicationDeploymentListener, dummyAppDescriptorFileBuilder.getId());
-    assertAppsDir(NONE, new String[] {dummyAppDescriptorFileBuilder.getId()}, true);
-
-    reset(applicationDeploymentListener);
-
-    File configFile = new File(appsDir + separator + dummyAppDescriptorFileBuilder.getDeployedPath(),
-                               getConfigFilePathWithinArtifact(MULE_CONFIG_XML_FILE));
-    configFile.setLastModified(configFile.lastModified() + FILE_TIMESTAMP_PRECISION_MILLIS);
-
-    assertApplicationRedeploymentSuccess(dummyAppDescriptorFileBuilder.getId());
-  }
-
-  @Test
-  public void redeploysExplodedAppAfterStartup() throws Exception {
-    startDeployment();
-
-    addExplodedAppFromBuilder(emptyAppFileBuilder);
-
-    assertApplicationDeploymentSuccess(applicationDeploymentListener, emptyAppFileBuilder.getId());
-    assertAppsDir(NONE, new String[] {emptyAppFileBuilder.getId()}, true);
-
-    reset(applicationDeploymentListener);
-
-    File configFile =
-        new File(appsDir + "/" + emptyAppFileBuilder.getDeployedPath(), getConfigFilePathWithinArtifact(MULE_CONFIG_XML_FILE));
-    assertThat("Configuration file does not exists", configFile.exists(), is(true));
-    assertThat("Could not update last updated time in configuration file",
-               configFile.setLastModified(configFile.lastModified() + FILE_TIMESTAMP_PRECISION_MILLIS), is(true));
-
-    assertApplicationRedeploymentSuccess(emptyAppFileBuilder.getId());
-  }
-
-  @Test
-  public void redeploysBrokenExplodedAppOnStartup() throws Exception {
-    addExplodedAppFromBuilder(incompleteAppFileBuilder);
-
-    startDeployment();
-
-    assertDeploymentFailure(applicationDeploymentListener, incompleteAppFileBuilder.getId());
-
-    // Maintains app dir created
-    assertAppsDir(NONE, new String[] {incompleteAppFileBuilder.getId()}, true);
-    assertArtifactIsRegisteredAsZombie(incompleteAppFileBuilder.getId(), deploymentService.getZombieApplications());
-
-    reset(applicationDeploymentListener);
-
-    final ReentrantLock lock = deploymentService.getLock();
-    lock.lock();
-    try {
-      File configFile = new File(appsDir + "/" + incompleteAppFileBuilder.getDeployedPath(),
-                                 getConfigFilePathWithinArtifact(MULE_CONFIG_XML_FILE));
-      assertThat(configFile.exists(), is(true));
-      configFile.setLastModified(configFile.lastModified() + FILE_TIMESTAMP_PRECISION_MILLIS);
-    } finally {
-      lock.unlock();
-    }
-
-    assertFailedApplicationRedeploymentFailure(applicationDeploymentListener, incompleteAppFileBuilder.getId());
-  }
-
-  @Test
-  public void redeploysBrokenExplodedAppAfterStartup() throws Exception {
-    startDeployment();
-
-    addExplodedAppFromBuilder(incompleteAppFileBuilder);
-
-    assertDeploymentFailure(applicationDeploymentListener, incompleteAppFileBuilder.getId());
-
-    // Maintains app dir created
-    assertAppsDir(NONE, new String[] {incompleteAppFileBuilder.getId()}, true);
-    assertArtifactIsRegisteredAsZombie(incompleteAppFileBuilder.getId(), deploymentService.getZombieApplications());
-
-    reset(applicationDeploymentListener);
-
-    File configFile = new File(appsDir + "/" + incompleteAppFileBuilder.getDeployedPath(),
-                               getConfigFilePathWithinArtifact(MULE_CONFIG_XML_FILE));
-    updateFileModifiedTime(configFile.lastModified(), configFile);
-
-    assertFailedApplicationRedeploymentFailure(applicationDeploymentListener, incompleteAppFileBuilder.getId());
-  }
-
-  @Test
-  public void redeploysInvalidExplodedAppAfterSuccessfulDeploymentOnStartup() throws Exception {
-    addExplodedAppFromBuilder(dummyAppDescriptorFileBuilder);
-
-    startDeployment();
-
-    assertApplicationDeploymentSuccess(applicationDeploymentListener, dummyAppDescriptorFileBuilder.getId());
-
-    assertAppsDir(NONE, new String[] {dummyAppDescriptorFileBuilder.getId()}, true);
-
-    reset(applicationDeploymentListener);
-
-    File originalConfigFile = new File(appsDir + "/" + dummyAppDescriptorFileBuilder.getDeployedPath(),
-                                       getConfigFilePathWithinArtifact(MULE_CONFIG_XML_FILE));
-    URL url = getClass().getResource(BROKEN_CONFIG_XML);
-    File newConfigFile = new File(url.toURI());
-    copyFile(newConfigFile, originalConfigFile);
-
-    assertApplicationRedeploymentFailure(dummyAppDescriptorFileBuilder.getId());
-  }
-
-  @Test
-  @Ignore("MULE-16403")
-  @Flaky
-  public void redeploysInvalidExplodedAppAfterSuccessfulDeploymentAfterStartup() throws Exception {
-    startDeployment();
-
-    addExplodedAppFromBuilder(dummyAppDescriptorFileBuilder);
-
-    assertApplicationDeploymentSuccess(applicationDeploymentListener, dummyAppDescriptorFileBuilder.getId());
-    assertAppsDir(NONE, new String[] {dummyAppDescriptorFileBuilder.getId()}, true);
-
-    reset(applicationDeploymentListener);
-
-    File originalConfigFile = new File(appsDir + "/" + dummyAppDescriptorFileBuilder.getDeployedPath(),
-                                       getConfigFilePathWithinArtifact(MULE_CONFIG_XML_FILE));
-    assertThat(originalConfigFile.exists(), is(true));
-    URL url = getClass().getResource(BROKEN_CONFIG_XML);
-    File newConfigFile = new File(url.toURI());
-    copyFile(newConfigFile, originalConfigFile);
-
-    assertApplicationRedeploymentFailure(dummyAppDescriptorFileBuilder.getId());
-  }
-
-  @Test
-  public void redeploysFixedAppAfterBrokenExplodedAppOnStartup() throws Exception {
-    addExplodedAppFromBuilder(incompleteAppFileBuilder);
-
-    startDeployment();
-
-    assertDeploymentFailure(applicationDeploymentListener, incompleteAppFileBuilder.getId());
-
-    reset(applicationDeploymentListener);
-
-    File originalConfigFile = new File(appsDir + "/" + incompleteAppFileBuilder.getDeployedPath(),
-                                       getConfigFilePathWithinArtifact(MULE_CONFIG_XML_FILE));
-    assertThat(originalConfigFile.exists(), is(true));
-    URL url = getClass().getResource(EMPTY_APP_CONFIG_XML);
-    File newConfigFile = new File(url.toURI());
-    copyFile(newConfigFile, originalConfigFile);
-    assertFailedApplicationRedeploymentSuccess(incompleteAppFileBuilder.getId());
-
-    addPackedAppFromBuilder(emptyAppFileBuilder);
-    assertApplicationDeploymentSuccess(applicationDeploymentListener, emptyAppFileBuilder.getId());
-
-    // Check that the failed application folder is still there
-    assertAppFolderIsMaintained(incompleteAppFileBuilder.getId());
-  }
-
-  @Test
-  public void redeploysFixedAppAfterBrokenExplodedAppAfterStartup() throws Exception {
-    startDeployment();
-
-    addExplodedAppFromBuilder(incompleteAppFileBuilder);
-    assertDeploymentFailure(applicationDeploymentListener, incompleteAppFileBuilder.getId());
-
-    reset(applicationDeploymentListener);
-
-    ReentrantLock deploymentLock = deploymentService.getLock();
-    deploymentLock.lock();
-    try {
-      File originalConfigFile = new File(appsDir + "/" + incompleteAppFileBuilder.getDeployedPath(),
-                                         getConfigFilePathWithinArtifact(MULE_CONFIG_XML_FILE));
-      URL url = getClass().getResource(EMPTY_DOMAIN_CONFIG_XML);
-      File newConfigFile = new File(url.toURI());
-      copyFile(newConfigFile, originalConfigFile);
-    } finally {
-      deploymentLock.unlock();
-    }
-
-    assertFailedApplicationRedeploymentSuccess(incompleteAppFileBuilder.getId());
-
-    addPackedAppFromBuilder(emptyAppFileBuilder);
-    assertApplicationDeploymentSuccess(applicationDeploymentListener, emptyAppFileBuilder.getId());
-
-    // Check that the failed application folder is still there
-    assertAppFolderIsMaintained(incompleteAppFileBuilder.getId());
-  }
-
-  @Test
-  public void redeploysZipAppOnConfigChanges() throws Exception {
-    addPackedAppFromBuilder(dummyAppDescriptorFileBuilder);
-
-    startDeployment();
-
-    assertApplicationDeploymentSuccess(applicationDeploymentListener, dummyAppDescriptorFileBuilder.getId());
-
-    assertAppsDir(NONE, new String[] {dummyAppDescriptorFileBuilder.getId()}, true);
-    assertEquals("Application has not been properly registered with Mule", 1, deploymentService.getApplications().size());
-
-    reset(applicationDeploymentListener);
-
-    File configFile = new File(appsDir + "/" + dummyAppDescriptorFileBuilder.getDeployedPath(),
-                               getConfigFilePathWithinArtifact(MULE_CONFIG_XML_FILE));
-    configFile.setLastModified(configFile.lastModified() + FILE_TIMESTAMP_PRECISION_MILLIS);
-
-    assertApplicationRedeploymentSuccess(dummyAppDescriptorFileBuilder.getId());
-    assertEquals("Application has not been properly registered with Mule", 1, deploymentService.getApplications().size());
-    assertAppsDir(NONE, new String[] {dummyAppDescriptorFileBuilder.getId()}, true);
   }
 
   @Test
@@ -942,25 +669,6 @@ public class ApplicationDeploymentTestCase extends AbstractApplicationDeployment
   }
 
   @Test
-  public void redeployedFailedAppAfterTouched() throws Exception {
-    addExplodedAppFromBuilder(emptyAppFileBuilder);
-
-    File appFolder = new File(appsDir.getPath(), emptyAppFileBuilder.getId());
-
-    File configFile = new File(appFolder, getConfigFilePathWithinArtifact(MULE_CONFIG_XML_FILE));
-    writeStringToFile(configFile, "you shall not pass");
-
-    startDeployment();
-    assertDeploymentFailure(applicationDeploymentListener, emptyAppFileBuilder.getId());
-    reset(applicationDeploymentListener);
-
-    URL url = getClass().getResource(EMPTY_DOMAIN_CONFIG_XML);
-    copyFile(new File(url.toURI()), configFile);
-
-    assertFailedApplicationRedeploymentSuccess(emptyAppFileBuilder.getId());
-  }
-
-  @Test
   public void receivesMuleContextDeploymentNotifications() throws Exception {
     // NOTE: need an integration test like this because DefaultMuleApplication
     // class cannot be unit tested.
@@ -1067,11 +775,6 @@ public class ApplicationDeploymentTestCase extends AbstractApplicationDeployment
   }
 
   @Test
-  public void zaraza() {
-
-  }
-
-  @Test
   public void undeploysAppCompletelyEvenOnDisposingException() throws Exception {
     addPackedAppFromBuilder(emptyAppFileBuilder);
 
@@ -1151,120 +854,6 @@ public class ApplicationDeploymentTestCase extends AbstractApplicationDeployment
   }
 
   @Test
-  public void redeploysZipAppAfterDeploymentErrorOnStartup() throws Exception {
-    addPackedAppFromBuilder(incompleteAppFileBuilder);
-
-    startDeployment();
-
-    assertDeploymentFailure(applicationDeploymentListener, incompleteAppFileBuilder.getId());
-
-    // Deploys another app to confirm that DeploymentService has execute the updater thread
-    addPackedAppFromBuilder(emptyAppFileBuilder);
-
-    assertApplicationDeploymentSuccess(applicationDeploymentListener, emptyAppFileBuilder.getId());
-
-    // Deploys another app to confirm that DeploymentService has execute the updater thread
-    addPackedAppFromBuilder(emptyAppFileBuilder, incompleteAppFileBuilder.getZipPath());
-    assertFailedApplicationRedeploymentSuccess(incompleteAppFileBuilder.getId());
-
-    assertNoZombiePresent(deploymentService.getZombieApplications());
-  }
-
-  @Test
-  public void redeploysZipAppAfterDeploymentErrorAfterStartup() throws Exception {
-    startDeployment();
-
-    addPackedAppFromBuilder(incompleteAppFileBuilder);
-
-    assertDeploymentFailure(applicationDeploymentListener, incompleteAppFileBuilder.getId());
-
-    // Deploys another app to confirm that DeploymentService has execute the updater thread
-    addPackedAppFromBuilder(emptyAppFileBuilder);
-    assertApplicationDeploymentSuccess(applicationDeploymentListener, emptyAppFileBuilder.getId());
-
-    // Deploys another app to confirm that DeploymentService has execute the updater thread
-    addPackedAppFromBuilder(emptyAppFileBuilder, incompleteAppFileBuilder.getZipPath());
-    assertFailedApplicationRedeploymentSuccess(incompleteAppFileBuilder.getId());
-
-    assertNoZombiePresent(deploymentService.getZombieApplications());
-  }
-
-  @Test
-  public void redeploysInvalidZipAppAfterSuccessfulDeploymentOnStartup() throws Exception {
-    addPackedAppFromBuilder(emptyAppFileBuilder);
-
-    startDeployment();
-
-    assertApplicationDeploymentSuccess(applicationDeploymentListener, emptyAppFileBuilder.getId());
-
-    reset(applicationDeploymentListener);
-
-    addPackedAppFromBuilder(incompleteAppFileBuilder, emptyAppFileBuilder.getZipPath());
-
-    assertApplicationRedeploymentFailure(emptyAppFileBuilder.getId());
-  }
-
-  @Test
-  public void redeploysInvalidZipAppAfterSuccessfulDeploymentAfterStartup() throws Exception {
-    startDeployment();
-
-    addPackedAppFromBuilder(emptyAppFileBuilder);
-    assertApplicationDeploymentSuccess(applicationDeploymentListener, emptyAppFileBuilder.getId());
-
-    addPackedAppFromBuilder(incompleteAppFileBuilder, emptyAppFileBuilder.getZipPath());
-    assertApplicationRedeploymentFailure(emptyAppFileBuilder.getId());
-  }
-
-  @Test
-  public void redeploysInvalidZipAppAfterFailedDeploymentOnStartup() throws Exception {
-    addPackedAppFromBuilder(incompleteAppFileBuilder);
-
-    startDeployment();
-
-    assertDeploymentFailure(applicationDeploymentListener, incompleteAppFileBuilder.getId());
-
-    reset(applicationDeploymentListener);
-
-    addPackedAppFromBuilder(incompleteAppFileBuilder);
-
-    assertFailedApplicationRedeploymentFailure(applicationDeploymentListener, incompleteAppFileBuilder.getId());
-  }
-
-  @Test
-  public void redeploysInvalidZipAppAfterFailedDeploymentAfterStartup() throws Exception {
-    startDeployment();
-
-    addPackedAppFromBuilder(incompleteAppFileBuilder);
-    assertDeploymentFailure(applicationDeploymentListener, incompleteAppFileBuilder.getId());
-
-    reset(applicationDeploymentListener);
-
-    addPackedAppFromBuilder(incompleteAppFileBuilder);
-
-    assertFailedApplicationRedeploymentFailure(applicationDeploymentListener, incompleteAppFileBuilder.getId());
-  }
-
-  @Test
-  public void redeploysExplodedAppAfterDeploymentError() throws Exception {
-    startDeployment();
-
-    addPackedAppFromBuilder(incompleteAppFileBuilder);
-
-    assertDeploymentFailure(applicationDeploymentListener, incompleteAppFileBuilder.getId());
-
-    // Deploys another app to confirm that DeploymentService has execute the updater thread
-    addPackedAppFromBuilder(emptyAppFileBuilder);
-
-    assertApplicationDeploymentSuccess(applicationDeploymentListener, emptyAppFileBuilder.getId());
-
-    // Redeploys a fixed version for incompleteApp
-    addExplodedAppFromBuilder(emptyAppFileBuilder, incompleteAppFileBuilder.getId());
-
-    assertFailedApplicationRedeploymentSuccess(incompleteAppFileBuilder.getId());
-    assertNoZombiePresent(deploymentService.getZombieApplications());
-  }
-
-  @Test
   public void deploysAppZipWithPlugin() throws Exception {
     final ApplicationFileBuilder echoPluginAppFileBuilder =
         appFileBuilder("dummyWithEchoPlugin").definedBy("app-with-echo-plugin-config.xml").dependingOn(echoPlugin);
@@ -1277,194 +866,9 @@ public class ApplicationDeploymentTestCase extends AbstractApplicationDeployment
   }
 
   @Test
-  public void deploysAppWithPluginSharedLibrary() throws Exception {
-    final ArtifactPluginFileBuilder echoPluginWithoutLib1 = new ArtifactPluginFileBuilder("echoPlugin1")
-        .configuredWith(EXPORTED_CLASS_PACKAGES_PROPERTY, "org.foo")
-        .containingClass(pluginEcho1TestClassFile, "org/foo/Plugin1Echo.class");
-
-    final ApplicationFileBuilder sharedLibPluginAppFileBuilder = appFileBuilder("shared-plugin-lib-app")
-        .definedBy("app-with-echo1-plugin-config.xml").dependingOn(echoPluginWithoutLib1)
-        .dependingOnSharedLibrary(new JarFileBuilder("barUtils", barUtils1_0JarFile));
-
-    addPackedAppFromBuilder(sharedLibPluginAppFileBuilder);
-
-    startDeployment();
-
-    assertApplicationDeploymentSuccess(applicationDeploymentListener, sharedLibPluginAppFileBuilder.getId());
-    assertAppsDir(NONE, new String[] {sharedLibPluginAppFileBuilder.getId()}, true);
-    assertApplicationAnchorFileExists(sharedLibPluginAppFileBuilder.getId());
-
-    executeApplicationFlow("main");
-  }
-
-  @Test
-  public void deploysAppWithPluginExportingAlreadyProvidedAppPackage() throws Exception {
-    // Defines a plugin that exports org.bar which is also exported on the application
-    ArtifactPluginFileBuilder echoPluginWithoutLib1 = new ArtifactPluginFileBuilder("echoPlugin1")
-        .configuredWith(EXPORTED_CLASS_PACKAGES_PROPERTY, "org.foo,org.bar")
-        .containingClass(pluginEcho1TestClassFile, "org/foo/Plugin1Echo.class");
-
-    ApplicationFileBuilder sharedLibPluginAppFileBuilder = appFileBuilder("shared-plugin-lib-app")
-        .definedBy("app-with-echo1-plugin-config.xml").dependingOn(echoPluginWithoutLib1)
-        .dependingOnSharedLibrary(new JarFileBuilder("barUtils", barUtils1_0JarFile));
-
-    addPackedAppFromBuilder(sharedLibPluginAppFileBuilder);
-
-    startDeployment();
-
-    assertApplicationDeploymentSuccess(applicationDeploymentListener, sharedLibPluginAppFileBuilder.getId());
-    assertAppsDir(NONE, new String[] {sharedLibPluginAppFileBuilder.getId()}, true);
-    assertApplicationAnchorFileExists(sharedLibPluginAppFileBuilder.getId());
-
-    executeApplicationFlow("main");
-  }
-
-  @Test
-  public void deploysAppWithExportedPackagePrecedenceOverPlugin() throws Exception {
-    // Defines a plugin that contains org.bar package, which is also exported on the application
-    ArtifactPluginFileBuilder echoPluginWithoutLib1 = new ArtifactPluginFileBuilder("echoPlugin1")
-        .configuredWith(EXPORTED_CLASS_PACKAGES_PROPERTY, "org.foo,org.bar")
-        .containingClass(pluginEcho1TestClassFile, "org/foo/Plugin1Echo.class")
-        .dependingOn(new JarFileBuilder("barUtils2_0", barUtils2_0JarFile));
-
-    ApplicationFileBuilder sharedLibPluginAppFileBuilder = appFileBuilder("shared-plugin-lib-app")
-        .definedBy("app-with-echo1-plugin-config.xml").dependingOn(echoPluginWithoutLib1)
-        .dependingOnSharedLibrary(new JarFileBuilder("barUtils", barUtils1_0JarFile));
-
-    addPackedAppFromBuilder(sharedLibPluginAppFileBuilder);
-
-    startDeployment();
-
-    assertApplicationDeploymentSuccess(applicationDeploymentListener, sharedLibPluginAppFileBuilder.getId());
-    assertAppsDir(NONE, new String[] {sharedLibPluginAppFileBuilder.getId()}, true);
-    assertApplicationAnchorFileExists(sharedLibPluginAppFileBuilder.getId());
-
-    executeApplicationFlow("main");
-  }
-
-  @Test
-  @Issue("MULE-17112")
-  @Description("If a plugin uses a library and the application sets another version of that library as a sharedLib, the plugin internally uses its own version of the lib and not the app's. "
-      + "Similar to deploysAppWithLibDifferentThanPlugin, but the bar2 dep in the app is shared in this case")
-  public void pluginWithDependencyAndConflictingVersionSharedByApp() throws Exception {
-    final ApplicationFileBuilder differentLibPluginAppFileBuilder = appFileBuilder("appWithLibDifferentThanPlugin")
-        .definedBy("app-plugin-different-lib-config.xml")
-        .dependingOn(echoPluginWithLib1)
-        .dependingOnSharedLibrary(new JarFileBuilder("barUtils2_0", barUtils2_0JarFile))
-        .containingClass(pluginEcho2TestClassFile, "org/foo/echo/Plugin2Echo.class");
-
-    addPackedAppFromBuilder(differentLibPluginAppFileBuilder);
-
-    startDeployment();
-
-    assertDeploymentSuccess(applicationDeploymentListener, differentLibPluginAppFileBuilder.getId());
-
-    executeApplicationFlow("main");
-  }
-
-  @Test
-  @Issue("MULE-17225")
-  public void appOverridingContainerClassAlsoPluginLocal() throws Exception {
-    File pluginEchoJavaxTestClassFile =
-        new SingleClassCompiler().dependingOn(barUtilsJavaxJarFile)
-            .compile(getResourceFile("/org/foo/echo/PluginJavaxEcho.java"));
-
-    ArtifactPluginFileBuilder echoPluginWithJavaxLib = new ArtifactPluginFileBuilder("echoPlugin1")
-        .configuredWith(EXPORTED_CLASS_PACKAGES_PROPERTY, "org.foo.echo")
-        .dependingOn(new JarFileBuilder("barUtilsJavax", barUtilsJavaxJarFile))
-        .containingClass(pluginEchoJavaxTestClassFile, "org/foo/echo/PluginJavaxEcho.class");
-
-    final ApplicationFileBuilder withJavaxEchoPlugin = appFileBuilder("appWithJavaxEchoPlugin")
-        .definedBy("app-with-javax-echo-plugin-config.xml")
-        .configuredWith(EXPORTED_CLASS_PACKAGES_PROPERTY, "javax.annotation")
-        .dependingOn(echoPluginWithJavaxLib)
-        .dependingOn(new JarFileBuilder("barUtilsJavaxB",
-                                        new JarCompiler().compiling(getResourceFile("/javax/annotation/BarUtils.java"))
-                                            .compile("bar-javax-b.jar")));
-
-    addPackedAppFromBuilder(withJavaxEchoPlugin);
-
-    startDeployment();
-
-    assertDeploymentSuccess(applicationDeploymentListener, withJavaxEchoPlugin.getId());
-
-    executeApplicationFlow("main");
-  }
-
-  @Test
-  @Description("Ensure that when a plugin2 depends on plugin1, when using something exported from plugin1, that is used")
-  public void pluginDependingAndExportingFromOtherPlugin() throws Exception {
-    ArtifactPluginFileBuilder echoPluginWithLib1 = new ArtifactPluginFileBuilder("echoPlugin1")
-        .configuredWith(EXPORTED_CLASS_PACKAGES_PROPERTY, "org.foo,org.bar")
-        .dependingOn(new JarFileBuilder("barUtils2", barUtils2_0JarFile))
-        .containingClass(pluginEcho1TestClassFile, "org/foo/Plugin1Echo.class");
-
-    // this plugin depends on a version of bar but will actually use the one from the dependant plugin
-    ArtifactPluginFileBuilder echoPluginWithLib2 = new ArtifactPluginFileBuilder("echoPlugin2")
-        // org.foo and org.bar are exported because plugin1 exports them.
-        // Similar to how the dependency beteween http and sockets work.
-        .configuredWith(EXPORTED_CLASS_PACKAGES_PROPERTY, "org.foo.echo,org.foo,org.bar")
-        .dependingOn(echoPluginWithLib1)
-        .containingClass(pluginEcho2TestClassFile, "org/foo/echo/Plugin2Echo.class");
-
-    final ApplicationFileBuilder usesPlugin2 = appFileBuilder("usesPlugin3")
-        .definedBy("app-with-echo2-plugin-config.xml")
-        .dependingOn(echoPluginWithLib2);
-
-    addPackedAppFromBuilder(usesPlugin2);
-
-    startDeployment();
-
-    assertDeploymentSuccess(applicationDeploymentListener, usesPlugin2.getId());
-
-    executeApplicationFlow("main");
-  }
-
-  @Test
-  @Description("Similar to pluginDependingAndExportingFromOtherPlugin, but for the case when there is no dependency between plugins. When plugin2 wants to use something local that is also exported by plugin1, the dependency local to plugin2 is used")
-  public void pluginNotDependingAndExportingFromOtherPlugin() throws Exception {
-    ArtifactPluginFileBuilder echoPluginWithLib1 = new ArtifactPluginFileBuilder("echoPlugin1")
-        .configuredWith(EXPORTED_CLASS_PACKAGES_PROPERTY, "org.foo,org.bar")
-        .dependingOn(new JarFileBuilder("barUtils1", barUtils1_0JarFile))
-        .containingClass(pluginEcho1TestClassFile, "org/foo/Plugin1Echo.class");
-
-    // this plugin depends on a version of bar but will actually use the one from the dependant plugin
-    ArtifactPluginFileBuilder echoPluginWithLib2 = new ArtifactPluginFileBuilder("echoPlugin2")
-        .configuredWith(EXPORTED_CLASS_PACKAGES_PROPERTY, "org.foo.echo")
-        .dependingOn(new JarFileBuilder("barUtils2", barUtils2_0JarFile))
-        .containingClass(pluginEcho2TestClassFile, "org/foo/echo/Plugin2Echo.class");
-
-    final ApplicationFileBuilder usesPlugin2 = appFileBuilder("usesPlugin3")
-        .definedBy("app-with-echo2-plugin-config.xml")
-        .dependingOn(echoPluginWithLib1)
-        .dependingOn(echoPluginWithLib2);
-
-    addPackedAppFromBuilder(usesPlugin2);
-
-    startDeployment();
-
-    assertDeploymentSuccess(applicationDeploymentListener, usesPlugin2.getId());
-
-    executeApplicationFlow("main");
-  }
-
-  @Test
   public void deploysAppZipWithExtensionPlugin() throws Exception {
     ApplicationFileBuilder applicationFileBuilder = createExtensionApplicationWithServices(APP_WITH_EXTENSION_PLUGIN_CONFIG,
                                                                                            helloExtensionV1Plugin);
-    addPackedAppFromBuilder(applicationFileBuilder);
-
-    startDeployment();
-
-    assertDeploymentSuccess(applicationDeploymentListener, applicationFileBuilder.getId());
-  }
-
-  @Test
-  public void deploysAppZipWithPrivilegedExtensionPlugin() throws Exception {
-    ArtifactPluginFileBuilder privilegedExtensionPlugin = createPrivilegedExtensionPlugin();
-
-    ApplicationFileBuilder applicationFileBuilder = appFileBuilder("privilegedPluginApp")
-        .definedBy(APP_WITH_PRIVILEGED_EXTENSION_PLUGIN_CONFIG).dependingOn(privilegedExtensionPlugin);
     addPackedAppFromBuilder(applicationFileBuilder);
 
     startDeployment();
@@ -1642,29 +1046,6 @@ public class ApplicationDeploymentTestCase extends AbstractApplicationDeployment
   }
 
   @Test
-  public void failsToDeployWithExtensionThatHasNonExistingIdForClassLoaderModel() throws Exception {
-    String extensionName = "extension-with-classloader-model-id-non-existing";
-
-    MulePluginModel.MulePluginModelBuilder builder =
-        new MulePluginModel.MulePluginModelBuilder().setName(extensionName).setMinMuleVersion(MIN_MULE_VERSION);
-    builder.withClassLoaderModelDescriptorLoader(new MuleArtifactLoaderDescriptorBuilder().setId("a-non-existing-ID-describer")
-        .addProperty("aProperty", "aValue").build());
-    builder.withBundleDescriptorLoader(createBundleDescriptorLoader(extensionName, MULE_EXTENSION_CLASSIFIER,
-                                                                    PROPERTIES_BUNDLE_DESCRIPTOR_LOADER_ID));
-
-    final ArtifactPluginFileBuilder byeXmlExtensionPlugin = new ArtifactPluginFileBuilder(extensionName)
-        .describedBy(builder.build());
-
-    ApplicationFileBuilder applicationFileBuilder = appFileBuilder("appWithExtensionXmlPluginFails")
-        .definedBy("app-with-extension-xml-plugin-module-bye.xml").dependingOn(byeXmlExtensionPlugin);
-    addPackedAppFromBuilder(applicationFileBuilder);
-
-    startDeployment();
-
-    assertDeploymentFailure(applicationDeploymentListener, applicationFileBuilder.getId(), times(1));
-  }
-
-  @Test
   public void deploysAppWithPluginBootstrapProperty() throws Exception {
     final ArtifactPluginFileBuilder pluginFileBuilder = new ArtifactPluginFileBuilder("bootstrapPlugin")
         .containingResource("plugin-bootstrap.properties", BOOTSTRAP_PROPERTIES)
@@ -1698,206 +1079,6 @@ public class ApplicationDeploymentTestCase extends AbstractApplicationDeployment
 
     assertDeploymentFailure(applicationDeploymentListener, applicationFileBuilder.getId());
   }
-
-  @Test
-  public void deploysMultiPluginVersionLib() throws Exception {
-    final ArtifactPluginFileBuilder echoPluginWithLib2 =
-        new ArtifactPluginFileBuilder("echoPlugin2").configuredWith(EXPORTED_CLASS_PACKAGES_PROPERTY, "org.foo.echo")
-            .dependingOn(new JarFileBuilder("barUtils2", barUtils2_0JarFile))
-            .containingClass(pluginEcho2TestClassFile, "org/foo/echo/Plugin2Echo.class");
-
-    final ApplicationFileBuilder multiLibPluginAppFileBuilder = appFileBuilder("multiPluginLibVersion")
-        .definedBy("multi-plugin-app-config.xml").dependingOn(echoPluginWithLib1).dependingOn(echoPluginWithLib2);
-
-    addPackedAppFromBuilder(multiLibPluginAppFileBuilder);
-
-    startDeployment();
-
-    assertDeploymentSuccess(applicationDeploymentListener, multiLibPluginAppFileBuilder.getId());
-
-    executeApplicationFlow("main");
-  }
-
-  @Test
-  public void deploysApplicationWithPluginDependingOnPlugin() throws Exception {
-
-    ArtifactPluginFileBuilder dependantPlugin =
-        new ArtifactPluginFileBuilder("dependantPlugin").configuredWith(EXPORTED_CLASS_PACKAGES_PROPERTY, "org.foo.echo")
-            .containingClass(pluginEcho3TestClassFile, "org/foo/echo/Plugin3Echo.class")
-            .dependingOn(echoPlugin);
-
-    final TestArtifactDescriptor artifactFileBuilder = appFileBuilder("plugin-depending-on-plugin-app")
-        .definedBy("plugin-depending-on-plugin-app-config.xml").dependingOn(dependantPlugin);
-    addPackedAppFromBuilder(artifactFileBuilder);
-
-    startDeployment();
-
-    assertDeploymentSuccess(applicationDeploymentListener, artifactFileBuilder.getId());
-
-    executeApplicationFlow("main");
-  }
-
-  @Test
-  public void deploysLightApplicationWithPluginDependingOnPlugin() throws Exception {
-
-    ArtifactPluginFileBuilder dependantPlugin =
-        new ArtifactPluginFileBuilder("dependantPlugin").configuredWith(EXPORTED_CLASS_PACKAGES_PROPERTY, "org.foo.echo")
-            .containingClass(pluginEcho3TestClassFile, "org/foo/echo/Plugin3Echo.class")
-            .dependingOn(echoPlugin);
-
-    File mavenRepoFolder = Paths.get(getMuleBaseFolder().getAbsolutePath(), "repository").toFile();
-    File testGroupIdRepoFolder = Paths.get(mavenRepoFolder.getAbsolutePath(), "org", "mule", "test").toFile();
-
-    JarFileBuilder testJarFileDependency = new JarFileBuilder("echoTestJar", echoTestJarFile);
-    copyFile(testJarFileDependency.getArtifactPomFile(),
-             Paths.get(testGroupIdRepoFolder.getAbsolutePath(), "echoTestJar", "1.0.0", "echoTestJar-1.0.0.pom").toFile());
-    copyFile(testJarFileDependency.getArtifactFile(),
-             Paths.get(testGroupIdRepoFolder.getAbsolutePath(), "echoTestJar", "1.0.0", "echoTestJar-1.0.0.jar").toFile());
-
-    copyFile(echoPlugin.getArtifactPomFile(),
-             Paths.get(testGroupIdRepoFolder.getAbsolutePath(), "echoPlugin", "1.0.0", "echoPlugin-1.0.0.pom").toFile());
-    copyFile(echoPlugin.getArtifactFile(),
-             Paths.get(testGroupIdRepoFolder.getAbsolutePath(), "echoPlugin", "1.0.0", "echoPlugin-1.0.0-mule-plugin.jar")
-                 .toFile());
-
-    copyFile(dependantPlugin.getArtifactPomFile(),
-             Paths.get(testGroupIdRepoFolder.getAbsolutePath(), "dependantPlugin", "1.0.0", "dependantPlugin-1.0.0.pom")
-                 .toFile());
-    copyFile(dependantPlugin.getArtifactFile(),
-             Paths.get(testGroupIdRepoFolder.getAbsolutePath(), "dependantPlugin", "1.0.0",
-                       "dependantPlugin-1.0.0-mule-plugin.jar")
-                 .toFile());
-
-    final TestArtifactDescriptor artifactFileBuilder = appFileBuilder("plugin-depending-on-plugin-app")
-        .definedBy("plugin-depending-on-plugin-app-config.xml").dependingOn(dependantPlugin).usingLightWeightPackage();
-    addPackedAppFromBuilder(artifactFileBuilder);
-
-    startDeployment();
-
-    assertDeploymentSuccess(applicationDeploymentListener, artifactFileBuilder.getId());
-
-    executeApplicationFlow("main");
-  }
-
-  @Test
-  public void deploysApplicationWithPrivilegedPluginDependingOnPlugin() throws Exception {
-    ArtifactPluginFileBuilder echoPlugin = new ArtifactPluginFileBuilder("echoPlugin")
-        .configuredWith(PRIVILEGED_EXPORTED_CLASS_PACKAGES_PROPERTY, "org.foo")
-        .configuredWith(PRIVILEGED_ARTIFACTS_PROPERTY, "org.mule.test:dependantPlugin")
-        .dependingOn(new JarFileBuilder("echoTestJar", echoTestJarFile));
-
-    ArtifactPluginFileBuilder dependantPlugin =
-        new ArtifactPluginFileBuilder("dependantPlugin").configuredWith(EXPORTED_CLASS_PACKAGES_PROPERTY, "org.foo.echo")
-            .containingClass(pluginEcho3TestClassFile, "org/foo/echo/Plugin3Echo.class")
-            .dependingOn(echoPlugin);
-
-    final TestArtifactDescriptor artifactFileBuilder = appFileBuilder("plugin-depending-on-plugin-app")
-        .definedBy("plugin-depending-on-plugin-app-config.xml").dependingOn(dependantPlugin);
-    addPackedAppFromBuilder(artifactFileBuilder);
-
-    startDeployment();
-
-    assertDeploymentSuccess(applicationDeploymentListener, artifactFileBuilder.getId());
-
-    executeApplicationFlow("main");
-  }
-
-  @Test
-  public void failsToDeployApplicationWithMissingPluginDependencyOnPlugin() throws Exception {
-
-    ArtifactPluginFileBuilder dependantPlugin =
-        new ArtifactPluginFileBuilder("dependantPlugin").configuredWith(EXPORTED_CLASS_PACKAGES_PROPERTY, "org.foo.echo")
-            .containingClass(pluginEcho3TestClassFile, "org/foo/echo/Plugin3Echo.class");
-
-    final TestArtifactDescriptor artifactFileBuilder = appFileBuilder("plugin-depending-on-plugin-app")
-        .definedBy("plugin-depending-on-plugin-app-config.xml").dependingOn(dependantPlugin);
-    addPackedAppFromBuilder(artifactFileBuilder);
-
-    startDeployment();
-
-    assertDeploymentSuccess(applicationDeploymentListener, artifactFileBuilder.getId());
-
-    try {
-      executeApplicationFlow("main");
-      fail("Expected to fail as there should be a missing class");
-    } catch (Exception e) {
-      assertThat(e.getCause().getCause(), instanceOf(MuleFatalException.class));
-      assertThat(e.getCause().getCause().getCause(), instanceOf(NoClassDefFoundError.class));
-      assertThat(e.getCause().getCause().getCause().getMessage(), containsString("org/foo/EchoTest"));
-    }
-  }
-
-  @Test
-  public void failsToDeployApplicationWithPluginDependantOnPluginNotShipped() throws Exception {
-    ArtifactPluginFileBuilder dependantPlugin =
-        new ArtifactPluginFileBuilder("dependantPlugin")
-            .dependingOn(echoPlugin);
-
-    final TestArtifactDescriptor artifactFileBuilder = appFileBuilder("plugin-depending-on-plugin-app")
-        .definedBy("plugin-depending-on-plugin-app-config.xml").dependingOn(dependantPlugin);
-    addPackedAppFromBuilder(artifactFileBuilder);
-
-    startDeployment();
-
-    assertDeploymentFailure(applicationDeploymentListener, artifactFileBuilder.getId(), times(1));
-  }
-
-  @Test
-  public void deploysAppWithLibDifferentThanPlugin() throws Exception {
-    final ApplicationFileBuilder differentLibPluginAppFileBuilder =
-        appFileBuilder("appWithLibDifferentThanPlugin").definedBy("app-plugin-different-lib-config.xml")
-            .dependingOn(echoPluginWithLib1).dependingOn(new JarFileBuilder("barUtils2_0", barUtils2_0JarFile))
-            .containingClass(pluginEcho2TestClassFile, "org/foo/echo/Plugin2Echo.class");
-
-    addPackedAppFromBuilder(differentLibPluginAppFileBuilder);
-
-    startDeployment();
-
-    assertDeploymentSuccess(applicationDeploymentListener, differentLibPluginAppFileBuilder.getId());
-
-    executeApplicationFlow("main");
-  }
-
-  @Test
-  public void deploysAppUsingPluginResource() throws Exception {
-    final ArtifactPluginFileBuilder pluginWithResource =
-        new ArtifactPluginFileBuilder("resourcePlugin").configuredWith(EXPORTED_RESOURCE_PROPERTY, "/pluginResource.properties")
-            .containingResource("pluginResourceSource.properties", "pluginResource.properties");
-
-    final ApplicationFileBuilder resourcePluginAppFileBuilder = appFileBuilder("dummyWithPluginResource")
-        .definedBy("plugin-resource-app-config.xml").dependingOn(pluginWithResource);
-
-    addPackedAppFromBuilder(resourcePluginAppFileBuilder);
-
-    startDeployment();
-
-    assertDeploymentSuccess(applicationDeploymentListener, resourcePluginAppFileBuilder.getId());
-  }
-
-  @Test
-  public void deploysAppProvidingResourceForPlugin() throws Exception {
-    File resourceConsumerClassFile =
-        new CompilerUtils.SingleClassCompiler().compile(getResourceFile("/org/foo/resource/ResourceConsumer.java"));
-
-    final ArtifactPluginFileBuilder pluginUsingAppResource =
-        new ArtifactPluginFileBuilder("appResourcePlugin").configuredWith(EXPORTED_CLASS_PACKAGES_PROPERTY, "org.foo.resource")
-            .containingClass(resourceConsumerClassFile, "org/foo/resource/ResourceConsumer.class");
-
-    final TestArtifactDescriptor artifactFileBuilder =
-        appFileBuilder("appProvidingResourceForPlugin")
-            .definedBy("app-providing-resource-for-plugin.xml")
-            .dependingOn(pluginUsingAppResource)
-            .configuredWith(EXPORTED_RESOURCES, "META-INF/app-resource.txt")
-            .usingResource(getResourceFile("/test-resource.txt").toString(), "META-INF/app-resource.txt");
-    addPackedAppFromBuilder(artifactFileBuilder);
-
-    startDeployment();
-
-    assertDeploymentSuccess(applicationDeploymentListener, artifactFileBuilder.getId());
-
-    executeApplicationFlow("main");
-  }
-
 
   @Test
   public void deploysMultipleAppsZipOnStartup() throws Exception {
@@ -2047,219 +1228,6 @@ public class ApplicationDeploymentTestCase extends AbstractApplicationDeployment
   }
 
   @Test
-  @Issue("MULE-16995")
-  @Description("This test covers an scenario where an app declares extensions-api as a shared library while using a privileged extension.")
-  public void appWithUnneededExtensionsApiDepDeploys() throws Exception {
-    String extensionsApiLib = getProperty("extensionsApiLib");
-    assumeThat(extensionsApiLib == null, is(false));
-
-    final ApplicationFileBuilder applicationFileBuilder = appFileBuilder("privilegedPluginApp")
-        .definedBy(APP_WITH_PRIVILEGED_EXTENSION_PLUGIN_CONFIG)
-        .dependingOn(createPrivilegedExtensionPlugin())
-        .dependingOnSharedLibrary(new JarFileBuilder("mule-extensions-api", new File(extensionsApiLib))
-            .withGroupId("org.mule.runtime")
-            .withVersion("1.1.6"));
-    addPackedAppFromBuilder(applicationFileBuilder);
-
-    startDeployment();
-    assertDeploymentSuccess(applicationDeploymentListener, applicationFileBuilder.getId());
-  }
-
-  @Test
-  public void appIncludingForbiddenJavaClass() throws Exception {
-    final ApplicationFileBuilder forbidden = appFileBuilder("forbidden")
-        .definedBy("app-with-forbidden-java-echo-plugin-config.xml")
-        .containingClass(pluginForbiddenJavaEchoTestClassFile, "org/foo/echo/PluginForbiddenJavaEcho.class")
-        .dependingOn(new JarFileBuilder("barUtilsForbiddenJavaJarFile", barUtilsForbiddenJavaJarFile));
-
-    addPackedAppFromBuilder(forbidden);
-
-    startDeployment();
-
-    assertDeploymentSuccess(applicationDeploymentListener, forbidden.getId());
-
-    try {
-      executeApplicationFlow("main");
-      fail("Expected to fail as there should be a missing class");
-    } catch (Exception e) {
-      assertThat(e.getCause().getCause(), instanceOf(MuleFatalException.class));
-      assertThat(e.getCause().getCause().getCause(), instanceOf(NoClassDefFoundError.class));
-      assertThat(e.getCause().getCause().getCause().getMessage(), containsString("java/lang/BarUtils"));
-    }
-  }
-
-  @Test
-  public void appIncludingForbiddenMuleContainerClass() throws Exception {
-    final ApplicationFileBuilder forbidden = appFileBuilder("forbidden")
-        .definedBy("app-with-forbidden-mule-echo-plugin-config.xml")
-        .containingClass(pluginForbiddenMuleContainerEchoTestClassFile, "org/foo/echo/PluginForbiddenMuleContainerEcho.class")
-        .dependingOn(new JarFileBuilder("barUtilsForbiddenMuleContainerJarFile", barUtilsForbiddenMuleContainerJarFile));
-
-    addPackedAppFromBuilder(forbidden);
-
-    startDeployment();
-
-    assertDeploymentSuccess(applicationDeploymentListener, forbidden.getId());
-
-    try {
-      executeApplicationFlow("main");
-      fail("Expected to fail as there should be a missing class");
-    } catch (Exception e) {
-      assertThat(e.getCause().getCause(), instanceOf(MuleFatalException.class));
-      assertThat(e.getCause().getCause().getCause(), instanceOf(NoClassDefFoundError.class));
-      assertThat(e.getCause().getCause().getCause().getMessage(), containsString("org/mule/runtime/api/util/BarUtils"));
-    }
-  }
-
-  @Test
-  public void appIncludingForbiddenMuleContainerThirdParty() throws Exception {
-    final ApplicationFileBuilder forbidden = appFileBuilder("forbidden")
-        .definedBy("app-with-forbidden-mule3rd-echo-plugin-config.xml")
-        .containingClass(pluginForbiddenMuleThirdPartyEchoTestClassFile, "org/foo/echo/PluginForbiddenMuleThirdPartyEcho.class")
-        .dependingOn(new JarFileBuilder("barUtilsForbiddenMuleThirdPartyJarFile", barUtilsForbiddenMuleThirdPartyJarFile));
-
-    addPackedAppFromBuilder(forbidden);
-
-    startDeployment();
-
-    assertDeploymentSuccess(applicationDeploymentListener, forbidden.getId());
-
-    try {
-      executeApplicationFlow("main");
-      fail("Expected to fail as there should be a missing class");
-    } catch (Exception e) {
-      assertThat(e.getCause().getCause(), instanceOf(MuleFatalException.class));
-      assertThat(e.getCause().getCause().getCause(), instanceOf(NoClassDefFoundError.class));
-      assertThat(e.getCause().getCause().getCause().getMessage(), containsString("org/slf4j/BarUtils"));
-    }
-  }
-
-  @Test
-  public void pluginIncludingForbiddenJavaClass() throws Exception {
-    ArtifactPluginFileBuilder echoPluginWithLib = new ArtifactPluginFileBuilder("echoPlugin2")
-        .configuredWith(EXPORTED_CLASS_PACKAGES_PROPERTY, "org.foo.echo")
-        .containingClass(pluginForbiddenJavaEchoTestClassFile, "org/foo/echo/PluginForbiddenJavaEcho.class")
-        .dependingOn(new JarFileBuilder("barUtilsForbiddenJavaJarFile", barUtilsForbiddenJavaJarFile));
-
-    final ApplicationFileBuilder usesPlugin2 = appFileBuilder("usesPlugin2")
-        .definedBy("app-with-forbidden-java-echo-plugin-config.xml")
-        .dependingOn(echoPluginWithLib);
-
-    addPackedAppFromBuilder(usesPlugin2);
-
-    startDeployment();
-
-    assertDeploymentSuccess(applicationDeploymentListener, usesPlugin2.getId());
-
-    try {
-      executeApplicationFlow("main");
-      fail("Expected to fail as there should be a missing class");
-    } catch (Exception e) {
-      assertThat(e.getCause().getCause(), instanceOf(MuleFatalException.class));
-      assertThat(e.getCause().getCause().getCause(), instanceOf(NoClassDefFoundError.class));
-      assertThat(e.getCause().getCause().getCause().getMessage(), containsString("java/lang/BarUtils"));
-    }
-  }
-
-  @Test
-  public void pluginIncludingForbiddenMuleContainerClass() throws Exception {
-    ArtifactPluginFileBuilder echoPluginWithLib = new ArtifactPluginFileBuilder("echoPlugin2")
-        .configuredWith(EXPORTED_CLASS_PACKAGES_PROPERTY, "org.foo.echo")
-        .containingClass(pluginForbiddenMuleContainerEchoTestClassFile, "org/foo/echo/PluginForbiddenMuleContainerEcho.class")
-        .dependingOn(new JarFileBuilder("barUtilsForbiddenMuleContainerJarFile", barUtilsForbiddenMuleContainerJarFile));
-
-    final ApplicationFileBuilder usesPlugin2 = appFileBuilder("usesPlugin2")
-        .definedBy("app-with-forbidden-mule-echo-plugin-config.xml")
-        .dependingOn(echoPluginWithLib);
-
-    addPackedAppFromBuilder(usesPlugin2);
-
-    startDeployment();
-
-    assertDeploymentSuccess(applicationDeploymentListener, usesPlugin2.getId());
-
-    try {
-      executeApplicationFlow("main");
-      fail("Expected to fail as there should be a missing class");
-    } catch (Exception e) {
-      assertThat(e.getCause().getCause(), instanceOf(MuleFatalException.class));
-      assertThat(e.getCause().getCause().getCause(), instanceOf(NoClassDefFoundError.class));
-      assertThat(e.getCause().getCause().getCause().getMessage(), containsString("org/mule/runtime/api/util/BarUtils"));
-    }
-  }
-
-  @Test
-  public void pluginIncludingForbiddenMuleContainerThirdParty() throws Exception {
-    ArtifactPluginFileBuilder echoPluginWithLib = new ArtifactPluginFileBuilder("echoPlugin2")
-        .configuredWith(EXPORTED_CLASS_PACKAGES_PROPERTY, "org.foo.echo")
-        .containingClass(pluginForbiddenMuleThirdPartyEchoTestClassFile, "org/foo/echo/PluginForbiddenMuleThirdPartyEcho.class")
-        .dependingOn(new JarFileBuilder("barUtilsForbiddenMuleThirdPartyJarFile", barUtilsForbiddenMuleThirdPartyJarFile));
-
-    final ApplicationFileBuilder usesPlugin2 = appFileBuilder("usesPlugin2")
-        .definedBy("app-with-forbidden-mule3rd-echo-plugin-config.xml")
-        .dependingOn(echoPluginWithLib);
-
-    addPackedAppFromBuilder(usesPlugin2);
-
-    startDeployment();
-
-    assertDeploymentSuccess(applicationDeploymentListener, usesPlugin2.getId());
-
-    try {
-      executeApplicationFlow("main");
-      fail("Expected to fail as there should be a missing class");
-    } catch (Exception e) {
-      assertThat(e.getCause().getCause(), instanceOf(MuleFatalException.class));
-      assertThat(e.getCause().getCause().getCause(), instanceOf(NoClassDefFoundError.class));
-      assertThat(e.getCause().getCause().getCause().getMessage(), containsString("org/slf4j/BarUtils"));
-    }
-  }
-
-  @Test
-  public void redeployMethodRedeploysIfApplicationIsAlreadyDeployedPacked() throws Exception {
-    DeploymentListener mockDeploymentListener = spy(new DeploymentStatusTracker());
-    deploymentService.addDeploymentListener(mockDeploymentListener);
-
-    // Deploy an application (packed)
-    addPackedAppFromBuilder(dummyAppDescriptorFileBuilder);
-    startDeployment();
-
-    // Application was deployed
-    assertApplicationDeploymentSuccess(applicationDeploymentListener, dummyAppDescriptorFileBuilder.getId());
-    verify(mockDeploymentListener, times(1)).onDeploymentSuccess(dummyAppDescriptorFileBuilder.getId());
-    verify(mockDeploymentListener, times(0)).onRedeploymentSuccess(dummyAppDescriptorFileBuilder.getId());
-
-    reset(mockDeploymentListener);
-
-    // Redeploy by using redeploy method
-    deploymentService.redeploy(dummyAppDescriptorFileBuilder.getArtifactFile().toURI());
-
-    // Application was redeployed
-    verify(mockDeploymentListener, times(1)).onRedeploymentSuccess(dummyAppDescriptorFileBuilder.getId());
-  }
-
-  @Test
-  public void redeployMethodRedeploysIfApplicationIsAlreadyDeployedExploded() throws Exception {
-    DeploymentListener mockDeploymentListener = spy(new DeploymentStatusTracker());
-    deploymentService.addDeploymentListener(mockDeploymentListener);
-
-    // Deploy an application (exploded)
-    addExplodedAppFromBuilder(dummyAppDescriptorFileBuilder);
-    startDeployment();
-
-    // Application was deployed
-    assertApplicationDeploymentSuccess(applicationDeploymentListener, dummyAppDescriptorFileBuilder.getId());
-    verify(mockDeploymentListener, times(1)).onDeploymentSuccess(dummyAppDescriptorFileBuilder.getId());
-    verify(mockDeploymentListener, times(0)).onRedeploymentSuccess(dummyAppDescriptorFileBuilder.getId());
-
-    // Redeploy by using redeploy method
-    deploymentService.redeploy(dummyAppDescriptorFileBuilder.getArtifactFile().toURI());
-
-    // Application was redeployed
-    verify(mockDeploymentListener, times(1)).onRedeploymentSuccess(dummyAppDescriptorFileBuilder.getId());
-  }
-
-  @Test
   public void deployMethodRedeploysIfApplicationIsAlreadyDeployedPacked() throws Exception {
     DeploymentListener mockDeploymentListener = spy(new DeploymentStatusTracker());
     deploymentService.addDeploymentListener(mockDeploymentListener);
@@ -2347,29 +1315,6 @@ public class ApplicationDeploymentTestCase extends AbstractApplicationDeployment
   private File getAppMetaFolder(Application app) {
     return new File((app.getRegistry().<MuleConfiguration>lookupByName(MuleProperties.OBJECT_MULE_CONFIGURATION).get()
         .getWorkingDirectory()));
-  }
-
-  private ArtifactPluginFileBuilder createPrivilegedExtensionPlugin() {
-    MulePluginModel.MulePluginModelBuilder mulePluginModelBuilder = new MulePluginModel.MulePluginModelBuilder()
-        .setMinMuleVersion(MIN_MULE_VERSION).setName(PRIVILEGED_EXTENSION_ARTIFACT_ID).setRequiredProduct(MULE)
-        .withBundleDescriptorLoader(createBundleDescriptorLoader(PRIVILEGED_EXTENSION_ARTIFACT_ID, MULE_EXTENSION_CLASSIFIER,
-                                                                 PROPERTIES_BUNDLE_DESCRIPTOR_LOADER_ID, "1.0.0"));
-    mulePluginModelBuilder.withClassLoaderModelDescriptorLoader(new MuleArtifactLoaderDescriptorBuilder()
-        .setId(MULE_LOADER_ID)
-        .build());
-    mulePluginModelBuilder.withExtensionModelDescriber().setId(JAVA_LOADER_ID)
-        .addProperty("type", "org.foo.hello.PrivilegedExtension")
-        .addProperty("version", "1.0");
-    return new ArtifactPluginFileBuilder(PRIVILEGED_EXTENSION_ARTIFACT_ID)
-        .dependingOn(new JarFileBuilder("privilegedExtensionV1", privilegedExtensionV1JarFile))
-        .describedBy(mulePluginModelBuilder.build());
-  }
-
-  @Override
-  protected Set<String> getPrivilegedArtifactIds() {
-    Set<String> privilegedArtifactIds = new HashSet<>();
-    privilegedArtifactIds.add(PRIVILEGED_EXTENSION_ARTIFACT_FULL_ID);
-    return privilegedArtifactIds;
   }
 
   private void deployAfterStartUp(ApplicationFileBuilder applicationFileBuilder) throws Exception {
