+++ modules/deployment/src/test/java/org/mule/runtime/module/deployment/internal/ApplicationPolicyDeploymentTestCase.java
@@ -31,6 +31,7 @@ import static org.mule.runtime.module.deployment.impl.internal.policy.Properties
 import static org.mule.runtime.module.deployment.internal.TestPolicyProcessor.invocationCount;
 import static org.mule.runtime.module.deployment.internal.TestPolicyProcessor.policyParametrization;
 import static org.mule.runtime.module.extension.api.loader.java.DefaultJavaExtensionModelLoader.JAVA_LOADER_ID;
+import static org.mule.test.allure.AllureConstants.ClassloadingIsolationFeature.CLASSLOADING_ISOLATION;
 
 import org.mule.runtime.api.deployment.meta.MuleArtifactLoaderDescriptor;
 import org.mule.runtime.api.deployment.meta.MuleArtifactLoaderDescriptorBuilder;
@@ -65,6 +66,7 @@ import org.junit.Ignore;
 import org.junit.Test;
 
 import io.qameta.allure.Description;
+import io.qameta.allure.Feature;
 import io.qameta.allure.Issue;
 
 /**
@@ -285,6 +287,7 @@ public class ApplicationPolicyDeploymentTestCase extends AbstractDeploymentTestC
 
 
   @Test
+  @Feature(CLASSLOADING_ISOLATION)
   public void appliesApplicationPolicyUsingAppPlugin() throws Exception {
     policyManager.registerPolicyTemplate(policyUsingAppPluginFileBuilder.getArtifactFile());
 
@@ -304,6 +307,7 @@ public class ApplicationPolicyDeploymentTestCase extends AbstractDeploymentTestC
   }
 
   @Test
+  @Feature(CLASSLOADING_ISOLATION)
   public void appliesApplicationPolicyUsingPluginOnlyInPolicy() throws Exception {
     policyManager.registerPolicyTemplate(policyIncludingPluginFileBuilder.getArtifactFile());
 
@@ -323,6 +327,7 @@ public class ApplicationPolicyDeploymentTestCase extends AbstractDeploymentTestC
   }
 
   @Test
+  @Feature(CLASSLOADING_ISOLATION)
   public void appliesApplicationPolicyIncludingPlugin() throws Exception {
     ArtifactPluginFileBuilder simpleExtensionPlugin = createSingleExtensionPlugin();
 
@@ -344,6 +349,7 @@ public class ApplicationPolicyDeploymentTestCase extends AbstractDeploymentTestC
   }
 
   @Test
+  @Feature(CLASSLOADING_ISOLATION)
   public void appliesApplicationPolicyDuplicatingExtensionPlugin() throws Exception {
     policyManager.registerPolicyTemplate(policyIncludingPluginFileBuilder.getArtifactFile());
 
@@ -362,6 +368,7 @@ public class ApplicationPolicyDeploymentTestCase extends AbstractDeploymentTestC
   }
 
   @Test
+  @Feature(CLASSLOADING_ISOLATION)
   public void appliesApplicationPolicyUsingModuleThatUsesPlugin() throws Exception {
     PolicyFileBuilder policyIncludingByePlugin = createPolicyIncludingByePlugin();
     policyManager.registerPolicyTemplate(policyIncludingByePlugin.getArtifactFile());
@@ -382,6 +389,7 @@ public class ApplicationPolicyDeploymentTestCase extends AbstractDeploymentTestC
   }
 
   @Test
+  @Feature(CLASSLOADING_ISOLATION)
   public void appliesApplicationPolicyUsingModuleThatUsesPluginDuplicatedInTheApplication() throws Exception {
     PolicyFileBuilder policyIncludingByePlugin = createPolicyIncludingByePlugin();
     policyManager.registerPolicyTemplate(policyIncludingByePlugin.getArtifactFile());
@@ -402,6 +410,7 @@ public class ApplicationPolicyDeploymentTestCase extends AbstractDeploymentTestC
   }
 
   @Test
+  @Feature(CLASSLOADING_ISOLATION)
   public void appliesApplicationPolicyDuplicatingPlugin() throws Exception {
 
     policyManager.registerPolicyTemplate(exceptionThrowingPluginImportingPolicyFileBuilder.getArtifactFile());
@@ -427,6 +436,7 @@ public class ApplicationPolicyDeploymentTestCase extends AbstractDeploymentTestC
   }
 
   @Test
+  @Feature(CLASSLOADING_ISOLATION)
   public void appliesApplicationPolicyDuplicatingPluginOnDomain() throws Exception {
 
     addPackedDomainFromBuilder(exceptionThrowingPluginImportingDomain);
@@ -452,8 +462,8 @@ public class ApplicationPolicyDeploymentTestCase extends AbstractDeploymentTestC
     }
   }
 
-  @Ignore("MULE-15842: fix once we support declaring share objects plugins in policies")
   @Test
+  @Ignore("MULE-15842: fix once we support declaring share objects plugins in policies")
   public void failsToApplyApplicationPolicyWithPluginVersionMismatch() throws Exception {
     policyManager.registerPolicyTemplate(policyIncludingHelloPluginV2FileBuilder.getArtifactFile());
 
@@ -500,6 +510,7 @@ public class ApplicationPolicyDeploymentTestCase extends AbstractDeploymentTestC
   }
 
   @Test
+  @Feature(CLASSLOADING_ISOLATION)
   public void appliesPolicyThatUsesPolicyClassOnExpression() throws Exception {
     ArtifactPluginFileBuilder simpleExtensionPlugin = createSingleExtensionPlugin();
 
@@ -574,6 +585,28 @@ public class ApplicationPolicyDeploymentTestCase extends AbstractDeploymentTestC
   }
 
   @Test
+  @Feature(CLASSLOADING_ISOLATION)
+  public void policyWithExtensionUsingObjectStore() throws Exception {
+    policyManager.registerPolicyTemplate(policyWithPluginUsingObjectStore().getArtifactFile());
+
+    ArtifactPluginFileBuilder simpleExtensionPlugin = createSingleExtensionPlugin();
+    ApplicationFileBuilder applicationFileBuilder = createExtensionApplicationWithServices(APP_WITH_SIMPLE_EXTENSION_CONFIG,
+                                                                                           simpleExtensionPlugin);
+
+    addPackedAppFromBuilder(applicationFileBuilder);
+
+    startDeployment();
+    assertApplicationDeploymentSuccess(applicationDeploymentListener, applicationFileBuilder.getId());
+
+    policyManager.addPolicy(applicationFileBuilder.getId(), policyWithPluginUsingObjectStore().getArtifactId(),
+                            new PolicyParametrization(BAR_POLICY_ID, s -> true, 1, emptyMap(),
+                                                      getResourceFile("/policy-using-object-store.xml"),
+                                                      emptyList()));
+
+    executeApplicationFlow("main");
+  }
+
+  @Test
   public void redeployPolicyWithSecurityManagerDefined() throws Exception {
     ArtifactPluginFileBuilder simpleExtensionPlugin = createSingleExtensionPlugin();
 
@@ -614,6 +647,18 @@ public class ApplicationPolicyDeploymentTestCase extends AbstractDeploymentTestC
         .build())
         .containingClass(echoTestClassFile, "org/foo/EchoTest.class")
         .dependingOn(helloExtensionV1Plugin);
+  }
+
+  private PolicyFileBuilder policyWithPluginUsingObjectStore() {
+    MulePolicyModel.MulePolicyModelBuilder mulePolicyModelBuilder = new MulePolicyModel.MulePolicyModelBuilder()
+        .setMinMuleVersion(MIN_MULE_VERSION).setName(BAZ_POLICY_NAME)
+        .setRequiredProduct(Product.MULE)
+        .withBundleDescriptorLoader(createBundleDescriptorLoader(BAZ_POLICY_NAME, MULE_POLICY_CLASSIFIER,
+                                                                 PROPERTIES_BUNDLE_DESCRIPTOR_LOADER_ID))
+        .withClassLoaderModelDescriptorLoader(new MuleArtifactLoaderDescriptor(MULE_LOADER_ID, emptyMap()));
+    return new PolicyFileBuilder(BAZ_POLICY_NAME).describedBy(mulePolicyModelBuilder
+        .build())
+        .dependingOn(usingObjectStorePlugin);
   }
 
   private PolicyFileBuilder createInjectedPolicy() throws URISyntaxException {
