+++ modules/deployment/src/test/resources/org/foo/os/UsingObjectStoreData.java
@@ -1,0 +1,32 @@
+/*
+ * Copyright (c) MuleSoft, Inc.  All rights reserved.  http://www.mulesoft.com
+ * The software in this package is published under the terms of the CPAL v1.0
+ * license, a copy of which has been included with this distribution in the
+ * LICENSE.txt file.
+ */
+
+package org.foo.os;
+
+import java.io.Serializable;
+
+import java.lang.String;
+
+import javax.inject.Inject;
+
+public class UsingObjectStoreData implements Serializable {
+
+  private String value;
+
+  public UsingObjectStoreData(String value) {
+    this.value = value;
+  }
+  
+  public String getValue() {
+    return value;
+  }
+  
+  public void setValue(String value) {
+    this.value = value;
+  }
+  
+}
