+++ modules/deployment/src/test/resources/org/foo/os/UsingObjectStoreExtension.java
@@ -1,0 +1,20 @@
+/*
+ * Copyright (c) MuleSoft, Inc.  All rights reserved.  http://www.mulesoft.com
+ * The software in this package is published under the terms of the CPAL v1.0
+ * license, a copy of which has been included with this distribution in the
+ * LICENSE.txt file.
+ */
+
+package org.foo.os;
+
+import org.mule.runtime.extension.api.annotation.Extension;
+import org.mule.runtime.extension.api.annotation.Operations;
+
+/**
+ * Extension for testing purposes
+ */
+@Extension(name = "Using ObjectStore")
+@Operations({UsingObjectStoreOperations.class})
+public class UsingObjectStoreExtension {
+
+}
