+++ modules/deployment/src/test/resources/policy-using-object-store.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<mule xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+      xmlns:test-policy="http://www.mulesoft.org/schema/mule/test-policy"
+      xmlns:using-object-store="http://www.mulesoft.org/schema/mule/using-object-store"
+      xmlns="http://www.mulesoft.org/schema/mule/core"
+      xsi:schemaLocation="http://www.mulesoft.org/schema/mule/core http://www.mulesoft.org/schema/mule/core/current/mule.xsd
+               http://www.mulesoft.org/schema/mule/test-policy http://www.mulesoft.org/schema/mule/test-policy/current/mule-test-policy.xsd
+               http://www.mulesoft.org/schema/mule/using-object-store http://www.mulesoft.org/schema/mule/using-object-store/current/mule-using-object-store.xsd">
+
+    <test-policy:proxy>
+        <test-policy:source>
+            <using-object-store:use-object-store/>
+        </test-policy:source>
+        <test-policy:operation>
+            <using-object-store:use-object-store/>
+        </test-policy:operation>
+    </test-policy:proxy>
+</mule>
